address mutianf review on #20099
- Rename `MonitoredResLabelKey*` → `MetricLabelKey*`; the tracer makes no
  resource vs metric distinction (relevant only to the Cloud Monitoring
  exporter, which now looks up the same keys via monitoredResLabelsSet).
- Introduce `DefaultMetricsProvider` so callers can opt into the built-in
  metrics explicitly instead of relying on the nil sentinel; nil still
  works for backward compatibility. Aliased as
  `bigtable.DefaultMetricsProvider`.
- `GenerateClientUID` falls back to "unknown" when `os.Hostname` fails
  instead of returning ("", err); the UUID alone still uniquely
  identifies the client.
- Delete unused `customExporter` field on `metricsConfig` (no production
  or test caller populated it); simplifies `newOtelMetricsContext`.
- Split lifecycle setup out of `tracer.go` into new `factory.go`
  (`Factory`, `NewFactory`, `NewFactoryForTest`, `builtInMeter…`,
  `NewAsyncRefreshErrHandler`, `createInstruments`, `CreateTracer`,
  `GenerateClientUID`, `CreateExporterOptions`, `disabledMetricsTracerFactory`).
  `tracer.go` now owns only runtime telemetry.
- Move `canonicalStatusStrings` + `CanonicalString` and
  `convertToGrpcStatusErr` from tracer.go / provider.go into util.go so
  helpers live alongside `GrpcCodeOf`. Move `methodNameReadRows` next to
  the first-response-latencies special case in tracer.go.
- `provider.go` shrinks to just the MetricsProvider interface and its
  two implementations.
- Drop dead `metricsPrefix` const; drop `maxAttrsLen` (compute the
  attribute slice capacity inline so adding a new metric can't leave the
  constant stale).
- accessors.go groups the four attribute methods together per mutianf.
